Fixing Settings.
Found out that the problem of it crashing.

On line 20 in SettingsFragments, I was getting an error on the logcat regarding the xml file. I realized that you are supposed to use android.support.v7.preference.SwitchPreferenceCompat other than SwitchPreference or else it crashes. This would go with the same with ListPreferences.